#30: replace ios assets compilation to IosMRGenerator
