+++ gradle-plugin/src/main/kotlin/dev/icerock/gradle/generator/MRGenerator.kt
@@ -34,6 +34,8 @@ abstract class MRGenerator(
         sourcesGenerationDir.deleteRecursively()
         resourcesGenerationDir.deleteRecursively()
 
+        beforeMRGeneration()
+
         @Suppress("SpreadOperator")
         val mrClassSpec = TypeSpec.objectBuilder(mrClassName)
             .addModifiers(*getMRClassModifiers())
@@ -61,6 +63,8 @@ abstract class MRGenerator(
 
         val file = fileSpec.build()
         file.writeTo(sourcesGenerationDir)
+
+        afterMRGeneration()
     }
 
     fun apply(project: Project) {
@@ -76,6 +80,9 @@ abstract class MRGenerator(
 
         apply(generationTask = genTask, project = project)
     }
+
+    protected open fun beforeMRGeneration() = Unit
+    protected open fun afterMRGeneration() = Unit
 
     protected abstract fun getMRClassModifiers(): Array<KModifier>
     protected abstract fun apply(generationTask: Task, project: Project)
